lw3/requirements.txt
pygame-ce~=2.4.1
PyTMX~=3.32